python-service/requirements.txt
selenium
requests
beautifulsoup4
grpcio
grpcio-tools
